requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
async-timeout==4.0.3
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
decorator==5.2.1
duckdb==1.4.1
exceptiongroup==1.3.0
executing==2.2.1
filetype==1.2.0
google-ai-generativelanguage==0.9.0
google-api-core==2.26.0
google-auth==2.41.1
googleapis-common-protos==1.71.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==8.37.0
jedi==0.19.2
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.79
langchain-google-genai==2.1.12
langchain-text-splitters==0.3.11
langgraph==0.6.10
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.37
matplotlib-inline==0.1.7
numpy==2.2.6
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sklearn==0.0
sniffio==1.3.1
SQLAlchemy==2.0.44
stack-data==0.6.3
tenacity==9.1.2
threadpoolctl==3.6.0
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
xxhash==3.6.0
zstandard==0.25.0
